PR Correction

+++ middleware/spi_ethernet/lib/src/spi_ethernet.c
@@ -104,11 +104,11 @@ int ethernet_send_frame( spi_ethernet_t *eth, ethernet_frame_t *frame ) {
     uint16_t pos = 0;
 
     // Destination MAC
-    for( int i = 0; i < 6; i++ )
+    for( uint8_t i = 0; i < 6; i++ )
         buffer[ pos++ ] = frame->dest[ i ];
 
     // Source MAC
-    for( int i = 0; i < 6; i++ )
+    for( uint8_t i = 0; i < 6; i++ )
         buffer[ pos++ ] = frame->src[ i ];
 
     // EtherType (big endian)
@@ -116,7 +116,7 @@ int ethernet_send_frame( spi_ethernet_t *eth, ethernet_frame_t *frame ) {
     buffer[ pos++ ] = ( frame->type & 0xFF );
 
     // Payload
-    for( int i = 0; i < frame->payload_len; i++ )
+    for( uint16_t i = 0; i < frame->payload_len; i++ )
         buffer[ pos++ ] = frame->payload[ i ];
 
     return spi_ethernet_send( eth, buffer, pos );
@@ -131,10 +131,10 @@ int ethernet_receive_frame( spi_ethernet_t *eth, ethernet_frame_t *frame ) {
 
     uint16_t pos = 0;
 
-    for( int i = 0; i < 6; i++ )
+    for( uint8_t i = 0; i < 6; i++ )
         frame->dest[ i ] = buffer[ pos++ ];
 
-    for( int i = 0; i < 6; i++ )
+    for( uint8_t i = 0; i < 6; i++ )
         frame->src[ i ] = buffer[ pos++ ];
 
     frame->type = ( buffer[ pos++ ] << 8 );
@@ -142,7 +142,7 @@ int ethernet_receive_frame( spi_ethernet_t *eth, ethernet_frame_t *frame ) {
 
     frame->payload_len = len - ETH_HEADER_SIZE;
 
-    for( int i = 0; i < frame->payload_len; i++ )
+    for( uint16_t i = 0; i < frame->payload_len; i++ )
         frame->payload[ i ] = buffer[ pos++ ];
 
     return frame->payload_len;
